MULTIARCH-4558: Set default for imagestreamImportMode based on observed config changes in the image config
As per https://github.com/openshift/enhancements/pull/1605, there was a
new field introduced in the image config spec and status which reflects
the global value to be set for imagestream import mode which is behind a
featuregate. This PR reads the observed config, checks if the importmode
string is present and uses it as the default when creating imagestreams.